requirements.txt
Flask
requests
python-dotenv
gunicorn


### **Adım 3: Instagram Fonksiyonlarını Ayırın (`instagram_helpers.py`)**



# instagram_helpers.py

import requests
import time
import os

# --- .env DOSYASINDAN BİLGİLERİ OKU ---
# Bu değerleri artık app.py'dan alacağız, ama test için burada kalabilirler
ACCESS_TOKEN = os.getenv("ACCESS_TOKEN")
INSTAGRAM_BUSINESS_ACCOUNT_ID = os.getenv("INSTAGRAM_BUSINESS_ACCOUNT_ID")
GRAPH_API_VERSION = os.getenv("GRAPH_API_VERSION")

# Bu dosyadaki fonksiyonlar aynı kalıyor, sadece __main__ bloğunu siliyoruz.

def create_reels_container(account_id, access_token, video_url, caption):
    """Adım 1: Reels'i yüklemek için bir medya konteyneri oluşturur."""
    print("Adım 1: Medya konteyneri oluşturuluyor...")
    # ... (fonksiyonun geri kalanı aynı) ...
    url = f"https://graph.facebook.com/{GRAPH_API_VERSION}/{account_id}/media"
    params = {
        'media_type': 'REELS',
        'video_url': video_url,
        'caption': caption,
        'share_to_feed': 'true',
        'access_token': access_token
    }
    try:
        response = requests.post(url, params=params)
        response.raise_for_status()
        result = response.json()
        if 'id' in result:
            print(f"Başarılı! Konteyner ID'si: {result['id']}")
            return result['id']
        else:
            print(f"Hata: Konteyner oluşturulamadı. Gelen yanıt: {result}")
            return None
    except requests.exceptions.RequestException as e:
        print(f"API isteği sırasında kritik bir hata oluştu: {e}")
        print(f"Detaylar: {e.response.json()}")
        return None

def check_container_status(creation_id, access_token):
    """Oluşturulan konteynerin durumunu perodik olarak kontrol eder."""
    # ... (fonksiyonun geri kalanı aynı) ...
    url = f"https://graph.facebook.com/{GRAPH_API_VERSION}/{creation_id}"
    params = {'fields': 'status_code', 'access_token': access_token}
    try:
        response = requests.get(url, params=params)
        response.raise_for_status()
        status = response.json().get('status_code')
        print(f"Konteyner durumu: {status}")
        return status
    except requests.exceptions.RequestException as e:
        print(f"Durum kontrolü sırasında hata: {e}")
        return "ERROR"


def publish_reels(account_id, creation_id, access_token):
    """Adım 2: Hazır olan konteyneri yayınlayarak Reels'i görünür hale getirir."""
    # ... (fonksiyonun geri kalanı aynı) ...
    print("\nAdım 2: Konteyner yayınlanıyor...")
    url = f"https://graph.facebook.com/{GRAPH_API_VERSION}/{account_id}/media_publish"
    params = {'creation_id': creation_id, 'access_token': access_token}
    try:
        response = requests.post(url, params=params)
        response.raise_for_status()
        result = response.json()
        if 'id' in result:
            print(f"🎉 TEBRİKLER! Reels başarıyla yayınlandı! Media ID: {result['id']}")
            return result['id']
        else:
            print(f"Hata: Reels yayınlanamadı. Gelen yanıt: {result}")
            return None
    except requests.exceptions.RequestException as e:
        print(f"Yayınlama sırasında kritik bir hata oluştu: {e}")
        print(f"Detaylar: {e.response.json()}")
        return None